Move dispersion fitter implementation to components
